feat(physics): fixed-timestep stepping with interpolation; clamp engine delta time
Variable-dt physics was the audit's top correctness finding: engine.rs
passed unclamped wall-clock deltas straight into Jolt, so one hitch
frame (or a backgrounded browser tab) fed dt=0.1+ into the solver —
tunneling, constraint explosions, non-determinism.

Native (physics_jolt.rs):
- per-world WorldStepState: accumulator over a fixed dt (default 1/60,
  setFixedTimestep), frame contribution clamped to 0.25s, catch-up
  capped at 4 steps/frame with backlog dropped — slows down instead of
  spiraling
- step_fixed returns alpha (remainder/fixed_dt); contact drain happens
  once per batch (the shim queue accumulates across sub-steps)
- opt-in render interpolation (set_interpolation): snapshot before the
  LAST step of a batch, position getter lerps, rotation getter nlerps
  with hemisphere correction; physics queries always see raw state
- state cleaned up on destroy_world

Web (jolt_bridge.js + new physics_ffi.rs): same semantics implemented
against JoltPhysics.js; web lib.rs physics surface split into
physics_ffi.rs (first slice of the >2000-line file ratchet).

Engine: delta_time clamped to MAX_DELTA_TIME=0.25 in begin_frame.

TS API: physics.step(world, dt) is now fixed-timestep by default and
returns alpha; stepVariable() is the explicit exact-dt opt-out;
setFixedTimestep / setInterpolation / getStepAlpha added. docs/physics.md
gains a Stepping section.

Tests: accumulator-vs-manual-stepping trajectory equivalence, hitch
clamp + backlog drop, interpolation blending (82 shared tests green).
